refactor(expense-documents): extract voidDocument into ExpenseDocumentLifecycleService (verbatim)
Phase 2c (final lifecycle slice) of transactional-core decompose. voidDocument
moved verbatim into the lifecycle service; facade delegates + sheds the journal
dep. Reversal JE + state transitions byte-identical (proven by the ~28 existing
void tests, unchanged). 436 tests green, tsc 0, lint 0.
